Moved drawing into view's draw method.
This is needed in Mojave, but, unfortunately, this now exposes the fact that we were just drawing into an existing buffer, which doesn't really work with the default double-buffering in the screensaver framework. I guess, the solution lies in drawing into an image and then copying that into the view in draw.